+++ ds2024Team77/src/main/java/gr/hua/dit/ds/ds2024Team77/repository/ReportRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ReportRepository extends JpaRepository<Report,Integer> {
+public interface ReportRepository extends JpaRepository<Report,Long> {
     List<Report> findByStatus(String status);
 }
