+++ ds2024Team77/src/main/java/gr/hua/dit/ds/ds2024Team77/repository/ReviewRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ReviewRepository extends JpaRepository<Review, Integer> {
-    Optional<Review> getByReviewee_Id(Integer revieweeId);
+public interface ReviewRepository extends JpaRepository<Review, Long> {
+    Optional<Review> getByReviewee_Id(Long revieweeId);
 
 }
